creating a fucking tester of my own

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vjean <vjean@student.42quebec.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/10 10:22:49 by vjean             #+#    #+#             */
-/*   Updated: 2022/06/29 16:37:08 by vjean            ###   ########.fr       */
+/*   Updated: 2022/06/29 20:18:26 by vjean            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,13 @@
 # include <stdlib.h>
 # include <stdarg.h>
 # include <stdio.h>
+# include <limits.h>
 
 void	ft_putchar(char c, int *len);
 void	ft_putnbr(int nb, int *len);
 void	ft_putstr(char *str, int *len);
-void	ft_puthex(unsigned long nb, char str, size_t base, int *len);
+void	ft_puthex(unsigned int nb, char str, size_t base, int *len);
 void	ft_puthex_p(unsigned long nb, char str, size_t base, int *len);
-int		ft_printf(const char *format, ...);
+int	ft_printf(const char *format, ...);
 
 #endif
